+++ validates.js
@@ -59,9 +59,9 @@ export default class Validates extends React.Component {
   }
 
   /**
-   * React lifecycle handler called immediately before the component's initial render
+   * React lifecycle handler called immediately after the component's initial render
    */
-  componentWillMount() {
+  componentDidMount() {
     // Update the handlers with the initial state
     this.onValidChange(this.props.validates);
   }